move db models into their files for better readability
