+++ anet_api/db/__init__.py
@@ -1,17 +1,1 @@
-from .models import (
-    Athlete,
-    AthleteRead,
-    AthleteCreate,
-    Meet,
-    MeetRead,
-    MeetCreate,
-    Team,
-    TeamRead,
-    TeamCreate,
-    Result,
-    ResultRead,
-    ResultCreate,
-    Race,
-    RaceCreate,
-    RaceRead,
-)
+from .models import *
