+++ anet_api/db/models/base.py
@@ -1,0 +1,8 @@
+from datetime import datetime
+
+from sqlmodel import Field, SQLModel
+
+
+class AbstractBase(SQLModel):
+    created_at: datetime = Field(default=datetime.utcnow(), nullable=False)
+    last_edited: datetime = Field(default_factory=datetime.utcnow, nullable=False)
